+++ app/src/main/java/org/piramalswasthya/sakhi/ui/home_activity/village_level_forms/maa_meeting/MaaMeetingListFragment.kt
@@ -38,7 +38,11 @@ class MaaMeetingListFragment : Fragment() {
         binding.btnNextPage.visibility = View.VISIBLE
 
         viewLifecycleOwner.lifecycleScope.launch {
-            binding.btnNextPage.isEnabled = !viewModel.hasMeetingInSameQuarter()
+            viewModel.maaMeetings.collect { meetings ->
+                val lastMeetingDate = meetings.lastOrNull()?.meetingDate
+                val hasMeeting = viewModel.hasMeetingInSameQuarter(lastMeetingDate)
+                binding.btnNextPage.isEnabled = !hasMeeting
+            }
         }
 
         binding.btnNextPage.setOnClickListener {
